examples/google_oauth2/requirements.txt
authlib==3.2.0
python-dotenv==1.0.0
